skills/zeelin-patent-retriever/requirements.txt
python-dotenv>=1.0.0
google-cloud-bigquery>=3.20.0
jsonschema>=4.20.0
